+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.1-dev.7
+- ✨ Add ability to fetch the success or failure values of an outcome without evaluating it
+- 🔥 Remove `isAsync` and `delayDuration`
+
 ## 0.0.1-dev.6
 - ✨ Add ability to only evaluate on a successful job